hacked update dns into little pieces and then made it a single repository not a workspace
