+++ Cargo.toml
@@ -1,8 +1,9 @@
-[workspace]
-members = [ "cloudflare", "config", "domain_services", "dyndns2", "ip_services", "requests","results", "update_dns_records"]
-resolver = "2"
+[package]
+name = "update_dns_records"
+version = "0.1.0"
+edition = "2021"
 
-[workspace.dependencies]
+[dependencies]
 serde_json = "1"
 bytes = "1"
 serde = { version = "1.0", features = ["derive"] }
