--- cloudflare/Cargo.toml
@@ -1,17 +1,0 @@
-[package]
-name = "cloudflare"
-version = "0.1.0"
-edition = "2021"
-
-# See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
-
-[dependencies]
-base64 = { workspace = true }
-bytes = { workspace = true }
-http-body-util = { workspace = true }
-hyper = { workspace = true }
-serde = { workspace = true }
-serde_json = { workspace = true }
-
-results = { path = "../results" }
-requests = { path = "../requests" }
