--- config/Cargo.toml
@@ -1,19 +1,0 @@
-[package]
-name = "config"
-version = "0.1.0"
-edition = "2021"
-
-# See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
-
-[dependencies]
-cloudflare = { path = "../cloudflare" }
-dyndns2 = { path = "../dyndns2" }
-ip_services = { path = "../ip_services" }
-serde = { workspace = true }
-serde_json = { workspace = true }
-tokio = { workspace = true }
-
-[features]
-default = ["cloudflare", "dyndns2"]
-dyndns2 = []
-cloudflare = []
