--- domain_services/Cargo.toml
@@ -1,17 +1,0 @@
-[package]
-name = "domain_services"
-version = "0.1.0"
-edition = "2021"
-
-# See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
-
-[dependencies]
-cloudflare = { path = "../cloudflare" }
-config = { path = "../config" }
-dyndns2 = { path = "../dyndns2" }
-results = { path = "../results" }
-
-[features]
-default = ["cloudflare", "dyndns2"]
-dyndns2 = []
-cloudflare = []
