--- dyndns2/Cargo.toml
@@ -1,16 +1,0 @@
-[package]
-name = "dyndns2"
-version = "0.1.0"
-edition = "2021"
-
-# See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
-
-[dependencies]
-base64 = { workspace = true }
-bytes = { workspace = true }
-http-body-util = { workspace = true }
-hyper = { workspace = true }
-requests = { path = "../requests" }
-results = { path = "../results" }
-serde = { workspace = true }
-serde_json = { workspace = true }
