--- ip_services/Cargo.toml
@@ -1,15 +1,0 @@
-[package]
-name = "ip_services"
-version = "0.1.0"
-edition = "2021"
-
-# See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
-
-[dependencies]
-bytes = { workspace = true }
-rand = "0.8.5"
-requests = { path = "../requests" }
-results = { path = "../results" }
-http-body-util = { workspace = true }
-hyper = { workspace = true }
-hyper-util = { workspace = true }
