+++ src/toolkit/Cargo.toml
@@ -1,9 +1,7 @@
 [package]
-name = "requests"
+name = "toolkit"
 version = "0.1.0"
-edition = "2021"
-
-# See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
+edition = "2024"
 
 [dependencies]
 bytes = { workspace = true }
@@ -11,7 +9,7 @@ http-body-util = { workspace = true }
 hyper = { workspace = true }
 hyper-util = { workspace = true }
 native-tls = { workspace = true }
-results = { path = "../results" }
-serde = { workspace = true }
 tokio = { workspace = true }
 tokio-native-tls = { workspace = true }
+serde = { workspace = true }
+serde_json = { workspace = true }
