+++ src/toolkit/ip_services.rs
@@ -1,0 +1,1 @@
+pub type IpServices = Vec<(String, String)>;
