--- update_dns_records/Cargo.toml
@@ -1,19 +1,0 @@
-[package]
-name = "update_dns_records"
-version = "0.1.0"
-edition = "2021"
-
-# See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
-
-[dependencies]
-bytes = { workspace = true }
-config = { path = "../config" }
-domain_services = { path = "../domain_services" }
-http-body-util = { workspace = true }
-hyper = { workspace = true }
-hyper-util = { workspace = true }
-ip_services = { path = "../ip_services" }
-native-tls = { workspace = true }
-results = { path = "../results" }
-tokio = { workspace = true }
-tokio-native-tls = { workspace = true }
